Flatten command.
Transforms gzipped data on stdin into "flat" data on stdout.

The flat data is the uncompressed file, plus a delta that
allows reconstructing the bit-identical compressed file.
This is stored into an uncompressed tar archive, following the
wrapped gz file format, with an extra entry, "uncompressed", containing
the uncompressed data.