Stop compressing as soon as streams don't match.
This is much more efficient when pristine-comp doesn't pick the
correct compressor on the first try.

kde-l10n-it-4.4.2.tar.bz2 is flattened in 35s instead of 69s.

Enabled only for bz2. gz looks at all compressed streams in xdelta
mode, and will either use a second pass for xdelta or SIGSTOP and
more complex pipe logic.